apply more lint, separate middlewares, more refactoring
