+++ middlewares/index.js
@@ -1,9 +1,10 @@
-import express from 'express';
 import bodyParser from 'body-parser';
-import { requestLogger } from './logger.js';
-import { errorResponder, errorLogger, invalidPathHandler } from './errors.js';
+import express from 'express';
 
-export function applyMiddlewares(app) {
+import { errorResponder, errorLogger, invalidPathHandler } from './errors.js';
+import requestLogger from './logger.js';
+
+export default function applyMiddlewares(app) {
   app.use('/uploads', express.static('uploads')); // access permission
   app.use(bodyParser.json());
   app.use(bodyParser.urlencoded({ extended: true }));
