+++ middlewares/logger.js
@@ -1,4 +1,4 @@
-export const requestLogger = (request, response, next) => {
+export default function (request, response, next) {
   console.log(`${request.method} url:: ${request.url}`);
   next();
-};
+}
